arreglado todo menos atoi

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: drosell- <drosell-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 15:25:44 by drosell-          #+#    #+#             */
-/*   Updated: 2022/09/20 10:33:09 by drosell-         ###   ########.fr       */
+/*   Updated: 2022/09/23 14:12:53 by drosell-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_atoi(const char *str)
 {
 	unsigned int	num;
 	int				i;
-	int				np;
+	long int		np;
 
 	np = 1;
 	i = 0;
